+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- fj-doc-maven-plugin, init goal, 'addJacoco' parameter <https://github.com/fugerit-org/fj-doc/issues/536>
+
+### Changed
+
+- fj-doc-maven-plugin, init goal, micronaut flavour version 4.9.4
+- fj-doc-maven-plugin, init goal, springboot-3 flavour version 3.5.6
+- fj-doc-val-pdfbox, pdfbox-version set to 2.0.35
+
 ## [8.16.8] - 2025-10-05
 
 ### Added